Fix threading issue by using signal-slot for requests too.
Instead of using a emitting a request signal connected to the request slot
of AsyncNetworkHandler, I was directly executing the user supplied
NetworkHandlers request method, which lead to some weird exceptions.

Everything was planned to be like this before but I forgot to make use of
it because initially (without lots of threading) it just seemed to work.